RF: Adjustments for clone move to core
