+++ datalad/distribution/install.py
@@ -49,7 +49,7 @@ from datalad.distribution.dataset import (
     EnsureDataset,
 )
 from datalad.distribution.get import Get
-from datalad.distribution.clone import Clone
+from datalad.core.distributed.clone import Clone
 
 __docformat__ = 'restructuredtext'
