+++ datalad/interface/__init__.py
@@ -78,7 +78,7 @@ _group_plumbing = (
     'Plumbing commands',
     [
         ('datalad.interface.annotate_paths', 'AnnotatePaths', 'annotate-paths'),
-        ('datalad.distribution.clone', 'Clone'),
+        ('datalad.core.distributed.clone', 'Clone'),
         ('datalad.distribution.create_test_dataset', 'CreateTestDataset',
          'create-test-dataset'),
         ('datalad.core.local.status', 'Status', 'status'),
